requirements.txt
# 网络请求
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# HTML 解析备选方案
# parsel>=1.8.0

# 日志
loguru>=0.7.0

# CLI 框架
typer[all]>=0.12.0
# rich 用于终端美化（typer 会依赖）
rich>=13.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 配置管理
python-dotenv>=1.0.0

# 日期时间处理
python-dateutil>=2.8.0
pytz>=2023.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0
